Migrate package structure to 'io.github.ramonster1' namespace.
Refactored all package declarations from 'com.jbeacon' to 'io.github.ramonster1' to align with the new project namespace. Updated corresponding imports and build.gradle group configuration to reflect this change. This is required for publishing Maven central using a GitHub namespace.
