+++ build.gradle
@@ -1,8 +1,9 @@
 plugins {
     id 'java-library'
+    id("eu.kakde.gradle.sonatype-maven-central-publisher") version "1.0.6"
 }
 
-group = 'com.jbeacon'
+group = 'io.github.ramonster1'
 version = '1.0-SNAPSHOT'
 
 repositories {
